// lib/operations/manual-outreach/build-manual-outreach-pack.ts
export type BuildManualOutreachPackDeps = {
  clinicRepo: ClinicRepository;
  crawlRepo: CrawlRepository;
  extractionRepo: ExtractionRepository;
  scoreRepo: ScoreRepository;
  outreachRepo: OutreachRepository;
  humanReviewRepo: HumanReviewRepository;
};